+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
 	"runtime/shiden",
 	"runtime/shibuya",
 	"tests/xcm-simulator",
+	"tests/integration",
 ]
 
 exclude = ["vendor"]
